refactor: highlightText 함수 shared/libs로 분리
